feat: RemoteTriggerMessageDialog を追加する
