+++ src/components/Dialog/RemoteTriggerMessageDialog/index.ts
@@ -1,0 +1,1 @@
+export { RemoteTriggerMessageDialog } from './RemoteTriggerMessageDialog'
