+++ src/components/Dialog/index.ts
@@ -7,4 +7,6 @@ export { DialogTrigger } from './DialogTrigger'
 export { DialogContent } from './DialogContent'
 export { DialogCloser } from './DialogCloser'
 export { ModelessDialog } from './ModelessDialog'
-export { RemoteDialogTrigger, RemoteTriggerActionDialog } from './RemoteTriggerActionDialog'
+export { RemoteTriggerActionDialog } from './RemoteTriggerActionDialog'
+export { RemoteTriggerMessageDialog } from './RemoteTriggerMessageDialog'
+export { RemoteDialogTrigger } from './RemoteDialogTrigger'
