+++ src/index.ts
@@ -36,6 +36,7 @@ export {
   ModelessDialog,
   RemoteDialogTrigger,
   RemoteTriggerActionDialog,
+  RemoteTriggerMessageDialog,
 } from './components/Dialog'
 export { Pagination } from './components/Pagination'
 export { RadioButton } from './components/RadioButton'
